+++ resources/lang/en/validation/endBorrowing.php
@@ -2,5 +2,6 @@
 
 return [
     'selectedBorrowings.required' => 'Select borrowings to end.',
-    'selectedBorrowings.*.distinct' => 'The borrowing of :item by :borrower was selected in multiple times.'
+    'selectedBorrowings.*.distinct' => 'The borrowing of :item by :borrower was selected in multiple times.',
+    'selectedBorrowings.*.no_self_return' => 'A user can\'t confirm the return of one of its own borrowings.'
 ];