+++ resources/lang/fr/validation/endBorrowing.php
@@ -2,5 +2,6 @@
 
 return [
     'selectedBorrowings.required' => 'Sélectionnez des emprunts à terminer.',
-    'selectedBorrowings.*.distinct' => 'L\'emprunt du jeu :item par :borrower a été choisi plusieurs fois.'
+    'selectedBorrowings.*.distinct' => 'L\'emprunt du jeu :item par :borrower a été choisi plusieurs fois.',
+    'selectedBorrowings.*.no_self_return' => 'Un utilisateur ne peut pas valider le retour de son propre emprunt.'
 ];